structure changes and correct table
