+++ src/components/UserTable/UserPageTable.tsx
@@ -1,29 +1,37 @@
 import { Table } from 'antd';
 
 const columns = [
-  { title: 'Name', dataIndex: 'name', key: 'name' },
-  { title: 'Age', dataIndex: 'age', key: 'age' },
-  { title: 'Address', dataIndex: 'address', key: 'address' },
+  { title: 'User', dataIndex: 'name', key: 'name' },
+  // { title: 'Age', dataIndex: 'age', key: 'age' },
+  // { title: 'Address', dataIndex: 'address', key: 'address' },
   {
     title: 'Action',
     dataIndex: '',
     key: 'x',
-    render: () => <a>Delete</a>
+    render: () => <a>Edit</a>,
+    edit: () => <a>Block</a>
+  },
+  {
+    title: '',
+    dataIndex: '',
+    key: 'x',
+    render: () => <a>Block | Unblock</a>
+    //edit: () => <a>Block</a>
   }
 ];
-
+// можно отсавить инфу о пользователях, и при нажатии на + будет выводить
 const data = [
   {
     key: 1,
-    name: 'John Brown',
-    age: 32,
-    address: 'New York No. 1 Lake Park',
+    name: 'Anna',
+    rest: 32,
+    position: 'New York No. 1 Lake Park',
     description:
       'My name is John Brown, I am 32 years old, living in New York No. 1 Lake Park.'
   },
   {
     key: 2,
-    name: 'Jim Green',
+    name: 'User 2',
     age: 42,
     address: 'London No. 1 Lake Park',
     description:
@@ -31,14 +39,14 @@ const data = [
   },
   {
     key: 3,
-    name: 'Not Expandable',
+    name: 'User 3',
     age: 29,
     address: 'Jiangsu No. 1 Lake Park',
     description: 'This not expandable'
   },
   {
     key: 4,
-    name: 'Joe Black',
+    name: 'User 4',
     age: 32,
     address: 'Sidney No. 1 Lake Park',
     description:
